requirements.txt
matplotlib
scikit-learn
tqdm
pillow
fastapi
uvicorn
python-multipart
